different method is implemented for guessing next interrupt moment

+++ Software/MPC_WS/MPC_Module2_dual_cla/headers_cla/cla_shared.h
@@ -33,10 +33,10 @@ extern float                M2_Iqref_cla;
 extern uint16_t             FaultFlagLocal;
 extern unsigned int         index_value_prev;
 extern float                M1_FswDecided_to_cla;
-extern float                M1_Interrupt_Moment_to_cla;
-extern float                M2_Interrupt_Moment_to_cla;
-extern float                M1_Next_Interrupt_Moment;
-extern float                M2_Next_Interrupt_Moment;
+extern int64_t                M1_Interrupt_Moment_to_cla;
+extern int64_t                M2_Interrupt_Moment_to_cla;
+extern int64_t                M1_Next_Interrupt_Moment;
+extern int64_t                M2_Next_Interrupt_Moment;
 extern float                fsw_cost_multipler;
 
 float CLA_Power(float base, uint16_t power);
